.../action/types/shared.ts,.*: add method to get attach. id to avoid repeating attachment fetching in card cover
* Update shared.ts's AttachmentActionBase to use getAttachmentId method
to fetch attachment ID, which is now simply overriden in
CardCoverActioBase to work for a different schema without the need to
repeat attachment fetching steps in fetchData, but call and reuse
fetchData on mixin.
* Update the type of propagated attachment key to allow it safely set to
undefined in final action types' Zod schemas, and asserted to never to
avoid accidentally accessing it.

Signed-off-by: Kiril Strezikozin <kirilstrezikozin@gmail.com>